requirements.txt
botbuilder-integration-aiohttp>=4.11.0
requests==2.28.2